+++ src/components/ChipsInput/ChipsInput.test.js
@@ -1,12 +1,13 @@
 import React from 'react'
-import { render, fireEvent } from '@testing-library/react'
+import { render, fireEvent, waitFor } from '@testing-library/react'
+import user from '@testing-library/user-event'
 import { screen } from '@testing-library/dom'
 import keymap from '../../utils/enums/keymap'
 import ChipsInput from './ChipsInput'
 
 const changeInput = value => {
   const input = screen.getByRole('textbox')
-  fireEvent.change(input, { target: { value } })
+  user.type(input, value)
   return input
 }
 
@@ -17,28 +18,32 @@ const addFreeTextChip = value => {
 
 describe('<ChipsInput />', () => {
   describe('chips creation', () => {
-    it('should fire onChange and onSubmit with relevant data when chip created', () => {
+    it('should fire onChange and onSubmit with relevant data when chip created', async () => {
       const onChange = jest.fn()
       const onSubmit = jest.fn()
       render(<ChipsInput onChange={onChange} onSubmit={onSubmit} />)
       addFreeTextChip('mockData')
-      addFreeTextChip('mockData2')
-      // 3 because onChange fire on first render
-      expect(onChange).toBeCalledTimes(3)
-      expect(onSubmit).toBeCalledTimes(2)
-      expect(onChange.mock.calls[2][0]).toEqual(['mockData', 'mockData2'])
-      expect(onSubmit.mock.calls[1][0]).toEqual(['mockData', 'mockData2'])
+      // 2 because onChange fire on first render
+      expect(onChange).toBeCalledTimes(2)
+      expect(onSubmit).toBeCalledTimes(1)
+      expect(onChange.mock.calls[1][0]).toEqual([
+        { label: 'mockData', value: 'mockData' },
+      ])
+      expect(onSubmit.mock.calls[0][0]).toEqual([
+        { label: 'mockData', value: 'mockData' },
+      ])
     })
 
-    it('should fire onInputChange with relevant data when chip created', () => {
+    it('should fire onInputChange with relevant data when chip created', async () => {
       const onInputChange = jest.fn()
       render(<ChipsInput onInputChange={onInputChange} />)
       changeInput('a')
       changeInput('b')
+      await waitFor(() => {})
       // 3 because onInputChange fire on first render + add & clear for each submit
       expect(onInputChange).toBeCalledTimes(3)
       expect(onInputChange.mock.calls[1][0]).toEqual('a')
-      expect(onInputChange.mock.calls[2][0]).toEqual('b')
+      expect(onInputChange.mock.calls[2][0]).toEqual('ab')
     })
 
     it('should fire renderChipIcon when submitting a chip', () => {
@@ -50,10 +55,12 @@ describe('<ChipsInput />', () => {
       expect(renderChipIcon.mock.calls[0][0]).toEqual({
         icon: undefined,
         label: 'mockData',
+        value: 'mockData',
       })
       expect(renderChipIcon.mock.calls[1][0]).toEqual({
         icon: undefined,
         label: 'mockData2',
+        value: 'mockData2',
       })
     })
 
@@ -74,6 +81,22 @@ describe('<ChipsInput />', () => {
       addFreeTextChip('1')
       const chipWithoutError = container.querySelector('.chipError')
       expect(chipWithoutError).toBe(null)
+    })
+  })
+
+  describe('autocomplete', () => {
+    it('should render autocomplete', async () => {
+      const autocomplete = async () => [{ label: 'first', value: '1' }]
+      const onChange = jest.fn()
+      render(<ChipsInput onChange={onChange} autocomplete={autocomplete} />)
+      const input = screen.getByRole('textbox')
+      fireEvent.focus(input)
+      await waitFor(() => screen.getByText('first'))
+      const firstAutocompleteItem = screen.getByText('first')
+      expect(firstAutocompleteItem).toBeTruthy()
+      user.click(firstAutocompleteItem)
+      await waitFor(() => screen.getByText('first'))
+      expect(screen.getByTestId('chip')).toBeTruthy()
     })
   })
 
@@ -125,14 +148,17 @@ describe('<ChipsInput />', () => {
 
     it('should remove chip when DELETE', () => {
       const onChange = jest.fn()
-      const { getAllByTestId, queryAllByTestId } = render(
+      const { getAllByTestId, queryAllByTestId, getByRole } = render(
         <ChipsInput onChange={onChange} />,
       )
       addFreeTextChip('mockData')
       addFreeTextChip('mockData2')
       const chips = getAllByTestId('chip')
+      const input = getByRole('textbox')
       const [firstChip] = chips
       fireEvent.click(firstChip)
+      fireEvent.keyDown(input, { keyCode: keymap.ARROW_LEFT })
+      fireEvent.keyDown(input, { keyCode: keymap.ARROW_LEFT })
       fireEvent.keyDown(firstChip, { keyCode: keymap.DELETE })
       fireEvent.keyDown(firstChip, { keyCode: keymap.DELETE })
       expect(onChange.mock.calls[4][0]).toEqual([])
@@ -161,8 +187,7 @@ describe('<ChipsInput />', () => {
       expect(document.activeElement).toBe(secondChip)
       fireEvent.keyDown(input, { keyCode: keymap.ARROW_RIGHT })
       expect(document.activeElement).toBe(input)
-      // Should be invoked once on mount and for each time focus entered and left the input
-      expect(onFocusChange).toBeCalledTimes(5)
+      expect(onFocusChange).toBeCalledTimes(3)
     })
   })